Renamed client subpackage to web
